requirements.txt
pandas==1.0.3
google-api-python-client==1.8.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
